requirements.txt
langchain-core==0.3.20
langchain-openai==0.2.6
python-dotenv==1.0.1
streamlit==1.37.1
openai==1.60.2
google-api-python-client==2.149.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
